Switch away from the deprecated IMG Graphics HAL APIs.
In DDK 1.4, a function table was provided at the end of the gralloc_module_t
to extend the gralloc API. This was replaced in DDK 1.5 with perform()
extensions, but the compatibility functions remained. Intel opted to
continue to use this compatibility path. However, future DDKs remove
the compatibility feature.

As DDK 1.5 has already been merged, it is possible to move away from the
deprecate extension methods without disruption.

This prepares the project for future DDKs.

Bug: 26592113
Change-Id: Ic99b449230fc994496d966e4c133a44a15b04125
Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>